resolved undefined symbol, but cuFFT plan does not support callback
The resolution is to have all callback stuff contained in cupy/cuda/cupy_cufft.cu,
so that at the linking stage for cupy_cufft_rdc.a the symbols are found. Then, at
the host linker I link with -lcufft so that other cuFFT API calls can be found.

It seems that cuFFT plan from the shared library simply doesn't support callback...